pyproject.toml
[project]
name = "release-api-client-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "digitalai-release-sdk>=25.3.1",
    "pydantic>=2.12.5",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests ",
    "integration: Integration tests with dev-environment",
]
testpaths = ["tests"]
python_files = ["test_*.py"]